refactor structure and add migration
