+++ src/main/java/com/sbh/bpm/JerseyConfig.java
@@ -2,6 +2,10 @@ package com.sbh.bpm;
 
 import javax.ws.rs.ApplicationPath;
 
+import com.sbh.bpm.controller.CustomController;
+import com.sbh.bpm.controller.GoogleCloudStorageController;
+import com.sbh.bpm.controller.NewBuildingController;
+
 import org.camunda.bpm.spring.boot.starter.rest.CamundaJerseyResourceConfig;
 import org.glassfish.jersey.media.multipart.MultiPartFeature;
 import org.springframework.stereotype.Component;
