+++ src/main/java/com/sbh/bpm/service/MultipartBody.java
@@ -1,4 +1,4 @@
-package com.sbh.bpm;
+package com.sbh.bpm.service;
 
 import java.io.InputStream;
